put `ends_with()` in different file
