+++ rootkit/Kbuild
@@ -1,1 +1,3 @@
 obj-m := yarin_module.o
+ccflags-y += -I$(src)/include
+yarin_module-y := yarin_module_main.o helpers.o
