+++ rootkit/include/helpers.h
@@ -1,0 +1,7 @@
+
+#ifndef YARIN_MODULE_HELPERS_H
+#define YARIN_MODULE_HELPERS_H
+
+int ends_with(const char *str, const char *suffix);
+
+#endif //YARIN_MODULE_HELPERS_H
